Change type of userWeight to int

+++ app/src/main/java/com/example/group22_hw02/SetWeightGender.java
@@ -18,7 +18,8 @@ import androidx.appcompat.app.AppCompatActivity;
 public class SetWeightGender extends AppCompatActivity {
     EditText weightWidget;
     RadioGroup genderGroup;
-    double userWeight;
+
+    int userWeight;
     double genderConstant;
     String gender;
